requirements_kjw.txt
aiofiles==23.2.1
albumentations==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.7
easydict==1.11
exceptiongroup==1.2.0
fastapi==0.108.0
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
fsspec==2023.12.2
ftfy==6.1.3
gradio==3.44.2
gradio_client==0.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
imutils==0.5.4
insightface==0.7.3
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.24.2
onnx==1.14.1
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==23.2
pandas==2.1.4
Pillow==10.0.1
prettytable==3.9.0
protobuf==4.23.2
psutil==5.9.5
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
pyvirtualcam==0.11.0
PyYAML==6.0.1
qudida==0.0.4
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.15.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
toolz==0.12.0
torch==2.0.1+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
wcwidth==0.2.12
websockets==11.0.3
zipp==3.17.0
